requirements.txt
blinker==1.7.0
Brotli==1.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
ConfigArgParse==1.7
coverage==7.4.1
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Login==0.6.3
gevent==24.2.1
geventhttpclient==2.0.11
greenlet==3.0.3
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
locust==2.23.1
MarkupSafe==2.1.5
msgpack==1.0.7
packaging==23.2
pluggy==1.4.0
psutil==5.9.8
pytest==8.0.1
pytest-cov==4.1.0
pytest-flask==1.3.0
python-dateutil==2.8.2
pyzmq==25.1.2
requests==2.31.0
roundrobin==0.0.4
six==1.16.0
urllib3==2.2.1
Werkzeug==3.0.1
zope.event==5.0
zope.interface==6.2
